Refactored show(), hide() and QShowEvent/QHideEvent.
Issues: show() is virtual, but was called many times. Using
hide and show events for managing widgets (dock windows, 
workspace, etc.) was hard, you needed a combination of ShowEvent,
ShowToParent, isHidden(), isVisible() and QEvent::spontaneous().

Solution: show() is only called once now. If you call it yourself, 
it will not be called again unless the widget is reparented.
ShowToParent and HideToParent are always sent when show() resp. hide()
are called.

The implementations of hide(), show(), showChildren() and hideChildren()
are much cleaner now. This was done to prepare a tighter integration of
QLayout and QWidget that possibly does not rely that much on posted
events.

[git-p4: depot-paths = "//depot/qt/main/": change = 102262]