First version of QOffset, for Large File Support in QFile API.
I expect QOffset to become QFile::Offset, and QIODevice::Offset
to become obsolete.

Note that this version allows conversion from uint or ulong to
QOffset, but not the reverse. Actually I'm not sure this is
the right thing, as users may need to access QOffset as POD,
for example for printing with pritnf. This is already the case
in Qt. I think I'll allow explicit conversion of QOffset to
an integer; this is less likely to break existing code. After
all this QOffset class is there mainly to allow writing and
reading to Q*Stream.

[git-p4: depot-paths = "//depot/qt/main/": change = 104218]