Ouch! Using a 64-bit integer internally results in crashes. I'm afraid that this will be extremely difficult to debug, unless this is the result a compiler bug: 	gcc version 3.2.2 20030222 (Red Hat Linux 3.2.2-5)
Back to a 32-bit integer for now.

I'll have to investigate that. The strange thing is that the
API of QOffset is still using 32-bit integers. So the 64-bit
integer within QOffset is cast to a 32-bit integer very early.
Could it be that casting 64-bit integers to 32-bit integers
doesn't work? Will look into that.

[git-p4: depot-paths = "//depot/qt/main/": change = 105177]