p4i integration Integrate 106074 from 3 to main: Integrate 106073 from 3.2 to 3: The qwinexport.h concept does not seem to make sense for Borland: with qwinexport.h, the compiler generates for every time a template is declared as export code and symbols; this results in a Qt DLL of size 11.5 MB oppposed to one of size 6.5 MB.
But the default behaviour of the compiler is to use the Smart
option: "When the Smart option is set, the compiler generates
public (global) definitions for all template instances. If more
than one module generates the same template instance, the linker
automatically merges duplicates to produce a single copy of the
instance."

So it exports the symbols for the templates and it does not do any
duplicates.

[git-p4: depot-paths = "//depot/qt/main/": change = 106075]