p4i integration Integrate 106161 from 3 to main: Integrate 106135 from 3.2 to 3:
* Problem
We press the "Page" button in the QToolBox, createing a
MouseButtonRelease event.
The designer mainwindow will intercept this event in the event filter,
and call selectionChanged() (that in turn will update the toolbar with
the layout toolbuttons using the layout in the current page).
The event is *then* processed by the toolbox and the page is changed.

* Fix
Make sure selectionChanged() is called after the MouseButtonRelease event
has been processed by QToolBox.

* Reviewed by
Espen and Matthias.

[git-p4: depot-paths = "//depot/qt/main/": change = 106163]