More compatibility for QLinkedList, and have the compatibility class QValueList inherit QLinkedList rather than QList.
This became an obvious step after the porting experiences Trenton,
Jasmin and Volker made. The C++ standard defines std::list to be a
linked list (iterators needs to remain valid when elements are inserted
and removed), QValueList implemented that behaviour, ergo we cannot
change it.

The only remaining porting issue is QStringList being a QList. This we
ease by providing a QValueList constructor that takes a QList, allowing
you to use a QValueList<QString> instead of QStringList.

[git-p4: depot-paths = "//depot/qt/main/": change = 106787]