p4i integration Integrate 106918 from 3 to main: Integrate 106915 from 3.2 to 3: Fix a bug in a new feature (QApplication::applicationDirPath() on Unix):
QFileInfo::readLink() returns for relative symlinks the relative
path (at least on my Linux machine). E.g. if I have an absolute
path /a/b/c, and b is a symlink to "bar", QFileReadLink() returns
"bar" without any slashes.

So if we replace "/b/" in "/a/b/c", we have to replace it by
"/bar/" instead of by symply "bar" (we need to get "/a/bar/c" and
not "/abarc").

Reviewed and approved by Espen.

[git-p4: depot-paths = "//depot/qt/main/": change = 106919]