p4i integration Integrate 107320 from 3 to main: Integrate 107318 from 3.2 to 3: fix regression found by adobe in task 22975: matching fixed pitch fonts is a bit special... most people do not care about the pitch of the font unless they explicitly request it.  for example, "Courier" is a fixed pitch font, but we do not want QFont::exactMatch() to return false if we didn't call setFixedPitch(TRUE).  however, we would want it to return false if we called setFixedPitch(FALSE).
we do this now by using an extra bit in QFontDef.  basically, we always
ignore the pitch when doing font matching and comparisons unless the
pitch was explicitly set.  this should be the behavior most people
want.

reviewed in detail by both matthias and espen.

note: i will probably have to fix this again in a few weeks :/

[git-p4: depot-paths = "//depot/qt/main/": change = 107321]