p4i integration Integrate 107410 from 3 to main: Integrate 107112 from 3.2 to 3:
Fix regression introduced with the previous full-width-selection
regression fix :(

Symptom was that when selecting paragraphs that spanned multiple lines
with linebreaks (pre-blocks), contents was chopped off rather randomly.

This code is unbelievably nasty, and this fix does not make it
better. If we continue hacking on this, then I cannot see how we will
be able to look our grandchildren straight in their eyes.

Reviewed by Brad.

[git-p4: depot-paths = "//depot/qt/main/": change = 107411]