p4i integration Integrate 107638 from 3 to main: Integrate 107637 from 3.2 to 3: some people work around the fact that QApplication sets the LC_NUMERIC locale to "C", which causes many things to break when using *printf()/*scanf() with float/double arguments.  Most notably, the postscript generator stops working and both QTextStream and QDataStream cannot stream float/double data that was created in another locale.
to work around this, we set LC_NUMERIC to the "C" locale in all places
where we use the *printf()/*scanf() family of functions (making sure to
reset the locale afterwards).

this fixes task 26562.  fix started by lars, finished by me, reviewed
by lars and me :)

note: the QString, QTextStream and QDataStream tests start working
after this change in locales which use a comma for a decimal separator.
Unfortunately, it is very difficult to determine exactly how to test
this automatically.

[git-p4: depot-paths = "//depot/qt/main/": change = 107639]