p4i integration Integrate 107838 from 3 to main: Integrate 107837 from 3.2 to 3: 2 fixes for font transformations:
- Xft 2.x supports bitmap font formats, but cannot transform them.  we
need to make sure that we do our standard bitmap transformation in this
case.

- Server side transformations do not seem to work correctly for all
types of fonts (for example, it works for bdf/pcf fonts, but not for
ttf).  It also seems to be extermely server dependent.  The best thing
is to just disable server side transformations until either server
support matures or we figure out a better way to do it.

reviewed by lars and espen.

[git-p4: depot-paths = "//depot/qt/main/": change = 107839]