p4i integration Integrate 107997 from 3 to main: Integrate 107941 from 3.2 to 3: Removed dndEnabled() test in contentsMousePressEvent for clearing selections to make KDE happy. They don't use the QListViewItem::dragEnabled (or setDragEnabled) so for them KListView brakes for multiple drag and drop.
This _reintroduces_ a small Extended selection bug. Select a range (with dnd disabled),
then select in the middle of the range a new range (using mouse move up and down). You will
see that the previous selection is not removed and you end up with old selections above
and below the new selectionrange still being there. If fixed, the clearing should be done
in mouseMoveEvent.

[git-p4: depot-paths = "//depot/qt/main/": change = 107998]