p4i integration Integrate 108482 from 3 to main: CE: Remove default show() to showMaximized(). The CE application developer should think about this, not the framework. QEmbedded does this too, but Qtopia doesn't.
Reorder the following functions:
   showWindow()
   showMinimized()
   showMaximized()
   showNormal()
and use big ifdef to clean up the code a bit. 
Also, change the implementation of these for CE, so we don't rely on the undocumented behaviour of SW_SHOWMAXIMIZED, which didn't work on some versions of CE. Now we reparent & move the toplevel window ourselvs, just like showFullscreen().

[git-p4: depot-paths = "//depot/qt/main/": change = 108485]