refactored painter-redirection. The painter keeps a stack now, allowing for apps to do QPainter::setRedirect(...) ... QPainter::restoreRedirected().
bitBlt now supports redirection as well. This made it possible to write
a REALLY nifty QPixmap::grabWidget() function. No more need to allocate
huge pixmaps for clipper widgets, or to allocate four pixmaps per child
widgets. Instead, the new function just has two: the result and a
buffer of the same size. Getting rid of the buffer would be extremely
tricky, it would require 100% reliable hard clipping (transformation
safe system clipping) on all platforms. I figured out this isn't worth
it.

Btw: QSharedDoubleBuffer is turned of during grabWidget(). The reason
is that it is a) not necessary to double buffer and b) the double
buffer will actually fail to provide proper backgrounds because of the
QSharedDoubleBuffer::begin(QPainter *p) function. In the redirect-case,
this painter \a p will point to a pixmap, and thus the double buffer
cannot figure out what backgound to pick depending on the device. Qt 4
will have something superior to QSharedDoubleBuffer.

[git-p4: depot-paths = "//depot/qt/main/": change = 108524]