p4i integration Integrate 108576 from 3 to main: Integrate 108575 from 3.2 to 3: Regression: Setting page size programmatically using QPrinter::setPageSize(PageSize) does not work on windows.
Reason: We introduced direct access to windows paper sizes (DMPAPER_xxx) using the
QPrinter::setWinPageSize() and QPrinter::winPageSize() functions for 3.2. This would
be done by setting the QPrinter page size to custom and using an extra value. Problem
is that we did not if pagesize was set to custom before writing the devmode, which 
makes setPageSize() useless.

Fix: use QPrinter::winPageSize() function instead, as this does the checking we
require, and it is tested.	

Tests: Tests for checking that setPageSize()/pageSize() and 
setWinPageSize()/winPageSize() exist. We now use a tested function with
intended behaviour.

Reviewed by: Lars and Marius2	

[git-p4: depot-paths = "//depot/qt/main/": change = 108595]