p4i integration Integrate 108837 from 3 to main: Integrate 108836 from 3.2 to 3: fix regression against 3.1.2
Description:
QDomNode::ownerDocument() should always return the ownerDocument if
there is one. Change 93016 unfortunately broke this in the case
that a node has a parent which has no parent/grandparent/etc. that
is the owner document.

In other words, the following case worked

    document
      +- level 1
          +- level 2

if you asked for the owner document, but the case

     level 1
      +- level 2

didn't work (even if level 1 has a owner document).

Task: 27424

Test: 108832

Review: Andreas

[git-p4: depot-paths = "//depot/qt/main/": change = 108838]