support double-buffering at the very core part of painting: in QWidget::repaint(). this is only done on X11 for the moment (and is a bit inefficient, but works as a proof of concept).  other platforms are to follow.
as a result of this, many things have happenend:

- WA_NoErase attribute renamed to WA_NoSystemBackground (see docs for
explanation).

- window system paint events end up calling QWidget::repaint()

- the new painter redirection didn't work correctly (since it would
actually try to modify the world matrix).  fix it to work properly and
to completely hide the redirection offset from the public API

- QWidget::erase() has been obsoleted in favor the automatic background
propagation/inheritance mechanism.  As a result, we will have
qt_erase_background functions for all platforms that simply do the
Right Thing(tm), instead of having each and every style doing the Wrong
Thing(tm).  (the styles included with Qt have been fixed as well).

- several old and new bugs that we noticed have been fixed (that were
directly related to background drawing)

[git-p4: depot-paths = "//depot/qt/main/": change = 108967]