p4i integration Integrate 109455 from 3 to main: Integrate 109454 from 3.2 to 3: Work around compiler bug/crash on Solaris with Sun Forte 6 update 1 compiler:
The destructor of a single QFontMetrics object would be called twice,
which is obviously wrong.  This is a compiler bug which we do not
totally understand yet, but it seems to be related to the ?: operator
and complex objects (those with constructors/destructors).

I was unable to reproduce this with Sun CC 5.5 (Sun ONE Studio 8), but
I have seen similar problems with the newer compiler (also related to
the ?: operator).  Unfortunately, I can't write a test case since I
cannot write a simple test to reproduce the problem.

Reviewed by Matthias and Lars.

See task 28040.

[git-p4: depot-paths = "//depot/qt/main/": change = 109456]