a few changes:
1) add copyright notices to all of the new qatomic* files that didn't
have one

2) using 'int volatile *' in q_cas_32 is the wrong thing to do, it
makes use of 'p' volatile, not when dereferencing (which is what we
want).  use 'volatile int *' instead.

3) valgrind complains about the current QAtomic::operator=(int)
function... the assignment operator is used throughout Qt to initialize
the QAtomic... this normally happens just after allocating some memory
on the heap for the QAtomic.  as such, valgrind would spit out a
warning about a conditional jump/move depending on an uninitialized
value.

the reason for this?  in the standard compare-and-set loop, 'expected'
would be initalized with something from the newly allocated (according
valgrind: uninitialized) memory, and when we could do a q_cas_32 which
would return the previous value (which is value from the newly
allocated memory).  We could compare the two together, see that they
were equal, and break out of the loop.  this works, is prefectly valid,
but confuses valgrind.

so, we change the semantics of QAtomic::operator= to work *only* as an
initializer (since that is all we use it for).  because of this, atomic
modifications to QAtomic can only be done with the ++ and -- operators.

[git-p4: depot-paths = "//depot/qt/main/": change = 109489]