p4i integration Integrate 109761 from 3 to main: manually integrate change 109758:
fix regression from 3.1: the initial size of a QMotifWidget will be
either what Qt thinks it should be, or what Motif thinks it should be.
however, the heuristic did not take into account that the programmer
could manually position the widget with QWidget::setGeometry() and
friends.  we do this now.

reviewed by Trond and Matthias.

see task 27380.

[git-p4: depot-paths = "//depot/qt/main/": change = 109762]