p4i integration Integrate 109921 from 3 to main: Integrate 109920 from 3.2 to 3: Regression: QPrinter crashes on windows 95, 98 and ME.
Reason: We use the function DeviceCapabilities to query the
printer for its bins. The bins are represented as an array of WORD's
We first call the function to determine the size of the array. Docs first state
that function return number of bytes required. Later that the number of
entries in the WORD array is returned. Wine docs confirm the latter. Since 
WORD is 16 bit and the array is only half the size when allocating bytes.

Fix: Allocate based on number of entries, not number of bytes.

Reviewed by Jazek

[git-p4: depot-paths = "//depot/qt/main/": change = 109922]