split up event loop into non gui and gui parts (currently on X11/embedded only).
This will allow us to split Qt a GUI lib and a kernel lib rather 
soon.

On Windows and Mac both eventloops are the same currently. On Win, this
is most probably the only way we can do it. On Mac, we should
probably use the regular unix implementation for the kernel lib,
and move qeventloop_mac.cpp completely over to the gui loop.

[git-p4: depot-paths = "//depot/qt/main/": change = 110298]