p4i integration Integrate 111026 from 3 to main: Integrate 111024 from 3.2 to 3: Regression from 3.1: if you xForm() a pixmap of size 20x20 with alpha channel with the matrix (34/20, 0, 0, 24/20, 0, 0), the resulting pixmap had size 34x34; in 3.2 it suddenly was 35x35 if (and only if) the QWMatrix::transformationMode was set to Areas.
Also a problem reported by a big customer (Adobe): when they draw a
pixmap, they sometimes get a qWarning() that the pixmap and the
mask are not of the same size. This is caused by the fact that the
mask and the pixmap were not calculated with the same code.

Task 29248.

Test added with change 111003 (and others).

Reviewed by Gunnar and Marius SO.

[git-p4: depot-paths = "//depot/qt/main/": change = 111027]