The type of a QVariant (i.e. the integer value) is used for streaming out QVariants over a QDataStream. So it is important that the type does not change (and if it change, we would probably need to have a mapping of the integer values in order to be able to read old versions). So assign the enum types concrete values; the values are taken from the order in Qt 3.2.
One remaining problem is that QCString maps now to 29 (ByteArray),
although it was 20 before (now unused). In order to read old
streamed files, we should maybe have a check somewhere.

Before this change, the QDom test crashed since it tried to read
some data from a "old" data streamed file, and suddenly the
ByteArray turned out to be a BitArray...

This change at least fixes that crash.

[git-p4: depot-paths = "//depot/qt/main/": change = 111307]