p4i integration Integrate 111717 from 3 to main: Integrate 111716 from 3.2 to 3: Crash fix on Windows 9x if you use alpha blended pixmaps with pixmap set to MemoryOptim.
The cause for the crash was that QPixmap::isMultiCellPixmap() said
TRUE for pixmaps with alpha channel, although they are not multi
cell pixmaps: The pixmap is first created as a multicell pixmap,
so the data->mcp is TRUE. So when we initialize the alpha pixmap,
we have to free the multi cell pixmap again by calling freeCell().

To avoid unnecessary allocation of a compatible bitmap, we call
freeCell() with TRUE. If we are doing this, we have to make sure
that the hdc (handle()) is set correctly since it might not be
initialized.

See also tasks 28161 and 29230.

Reviewed by Marius SO

[git-p4: depot-paths = "//depot/qt/main/": change = 111718]