p4i integration Integrate 113395 from 3 to main: Integrate 113394 from 3.2 to 3: fix a crash in QRichText when zooming.
Problem was that after zooming, two formats could have the same key in
the format collection. When the refcount of one of them dropped to zero,
the first format in the collection with the same key was deleted, which in
this case was not always equal to the format we wanted to delete.

See task #29160

Reviewed by Brad and Gunnar and checked with Valgrind

[git-p4: depot-paths = "//depot/qt/main/": change = 113396]