Once again... Compile on windows...
This time, keep QImage in the kernel library. This is basically 
messed up, and we'll split QImage into QKernelImage without mime/pixmap
dependencies.

The QVariant implemented in both kernel and gui library also does not
work since the pixmap related functions are not inlined properly on
MSVC.net. We'll split QVariant into QKernelVariant and QVariant to 
work around this. That should also get rid of the poor mans vft in
QVariant.

Finally. The Q_XXX_EXPORT logic was broken. We need to export only
when building each separate library. Include otherwise. Also introduced
the QT_BUILD_XXX_LIB to aid in this process.

Cheers Sam ;-)

[git-p4: depot-paths = "//depot/qt/main/": change = 113592]