p4i integration Integrate 113874 from 3 to main: new API introduction: QWidget::setFullScreen(bool)        ::setMinimized(bool)        ::setMaximized(bool)
These allow individual setting of the minimized/maximized/fullScreen
properties.  Fully implemented on X11 (with EWMH or legacy code).

The semantics of these functions are fairly straight forward.  The
existing showMinimized(), showMaximized(), showFullScreen() and
showNormal() functions have been reimplemented in terms of the above
functions.  They do not give the same kind of flexibility, though.  We
are considering obsoleting them in a future version.

Note: only have skeleton implementations on other platforms for the
moment

[git-p4: depot-paths = "//depot/qt/main/": change = 113974]