Compile with Borland: I don't understand what the problem actually is, but it seems to be a compiler bug; the culprit is related to the lines
    Q_DECLARE_OPERATORS_FOR_FLAGS(Qt::WidgetState);
    Q_DECLARE_OPERATORS_FOR_FLAGS(Qt::WindowFlags);
    Q_DECLARE_OPERATORS_FOR_FLAGS(Qt::StringComparisonMode);

in qnamespace.h. So I am able to reprodce the compile problem with
the following small code example:

    enum MyEnum;

    #if 1
    MyEnum operator|(MyEnum, MyEnum )
    { return (MyEnum)0;}
    #endif

    #include <windows.h>

    typedef DECLSPEC_IMPORT BOOL (WINAPI *PtrLookupAccountSidW)(LPCWSTR, PSID, LPWSTR, LPDWORD, LPWSTR, LPDWORD, PSID_NAME_USE);
    static PtrLookupAccountSidW ptrLookupAccountSidW = 0;

    void hmpfl()
    {
	PSID pOwner = 0;
	DWORD lowner = 0, ldomain = 0;
	SID_NAME_USE foobarBlubber;
	ptrLookupAccountSidW( NULL, pOwner, NULL, &lowner, NULL, &ldomain, &foobarBlubber );
    }

If the operator|() is there, it Borland (both 5 and 6) give me the
errors:

    Error E2034 example.cpp 18: Cannot convert 'int *' to '_SID_NAME_USE *' in function hmpfl()
    Error E2340 example.cpp 18: Type mismatch in parameter 7 (wanted '_SID_NAME_USE*', got 'int *') in function hmpfl()

Not defining the operator makes it compile. The concrete type of
the operator does not matter (I could reproduce the problem with
operator+() and operator++() as well). Changing the "enum MyEnum"
to "class MyEnum" on the other hand makes it work again.

Maybe Borland has a problem with operators defined on enum types?

Anyway, for now I don't think it is worth investigating more and we
have to live with the ugly cast.

[git-p4: depot-paths = "//depot/qt/main/": change = 114252]