p4i integration Integrate 114313 from 3 to main: MinGW build fix.
Please note that the file separator should be / instead of \ on
this platform. There are two reasons for this, and both have to
do with MSYS interacting with MinGW. It's a good idea to check
Qt builds with MSYS installed. It's even a good idea to check
that it builds from the MSYS shell instead of COMMAND.COM, because
those who use MinGW *will* use MSYS as well.
1) Older MSYS packages would apparently replace MinGW's original
mingw32-make with their own mingw32-make that wouldn't understand
the \ file separator. Dependencies couldn't work. It seems this
is not the case anymore with MinGW 3.10 and MSYS 1.09.
2) If MSYS' shell is in the current PATH, mingw32-make will prefer
it over COMMAND.COM. The MSYS shell doesn't understand \ and for
example thsi won't work:
	cd dir\subdir
while this will work with both the MSYS shell and COMMAND.COM:
	cd dir/subdir

[git-p4: depot-paths = "//depot/qt/main/": change = 114314]