p4i integration Integrate 115255 from 3 to main: Fixed printing of images, scaled using the draw functions taking a rect, on a rotated/sheared painter.
What was broken was that the scaling was applied after the transformation
leading to shearing and wrong translation. Applying scaling of the image
before translation (pre multiply the matrix) solved this. We then have 
to use the scaled matrix in the trueMatrix call and the size of the
original image to get the proper translation in the modified painter
matrix below.	

[git-p4: depot-paths = "//depot/qt/main/": change = 115256]