Removed all template exports from our code.
After some testing on msvc .net and 6.0 it turns out that the compiler 
warning C4251 "Class 'A' must have dll interface to be used of clients
of class 'B'", is not a problem. The warning occurs when a template 
specialization is used as a member variable of an exported class. In the
testcase I could not find any problems when all template exports where
removed, nor when building Qt without with .net. Lets see how this turns
out.	

The warning is currently disabled with a pragma to make the compile output
a bit nicer.

[git-p4: depot-paths = "//depot/qt/main/": change = 115520]