p4i integration Integrate 116791 from 3 to main: Integrate 116790 from 3.2 to 3: Fix gcc 3.4 compile problems and make our code comply to the C++ standard.
Problem was that only as many bits in the enum as needed to
store it's values are defined. Casts of integers outside
this range to the enums are undefined by the standard.

The case here was:

enum Foo { A, B, C }; Foo f = (Foo)-1; 

This lead to f actually being 3 when cast back to an integer.

See task #33342

Reviewed by Brad and Paul.

[git-p4: depot-paths = "//depot/qt/main/": change = 116792]