p4i integration Integrate 117174 from 3 to main: Fix a bug in incremental parsing: it could happen that some attributes occur more than once in the reporting event.
The reason for this is if the parsing runs exactly then out of data
when the parsing of an attribute is finished: we are then still in
the state Attribute so we call processAttribute(). But after that
we notice that we ran out of data, so we store our current state on
the parsing stack. When we pop the stack again, we are still in the
Attribute state so we call processAttribute() again, and as a
result the same attribute is added twice!

The solution to this is to split the Attribute state into two
states, Attrib and AttribPro depending if we already called
processAttribute() or not.

Test added with 117144 (some other things still fail in the test).

This fixes task 32996.

[git-p4: depot-paths = "//depot/qt/main/": change = 117175]