p4i integration Integrate 117368 from 3 to main: Integrate 115832 from 3CE to 3: Make QComboBox's autocompletion able to autocomplete on Virtual Keyboard actions. (Like the SIP on Windows CE)
Most virtual keyboards work in the way that they send a key press and a key release right after each other.

QComboBox does not like this, since it expects to get other events in between, to update the autocompletion state (typically a QPaintEvent, as a result of QLineEdit repainting after the key press event from the HW keyboard)

Seems that Arnt was overly careful when implementing this, and didn't permit any autocompletion to happen after the Key Release. So, remove this limitation, and now always stop autocompleting, even if it didn't result to anything.

[git-p4: depot-paths = "//depot/qt/main/": change = 117369]