p4i integration Integrate 118630 from 3 to main: Task: 26578 Reviewed by: Rainer
HTTP 1.1 can return a "100 Continue" header before the server sends the 
       data (see also sections 8.2.3 and 10.1.1 of RFC 2616). The QHttp
       implementation handled this as a the final response which is not
correct.

As the "100 Continue" header is sent by the server to confirm that
the client can continue posting the request body, and as QHttp always
posts the header and body in one chunk, QHttp now simply ignores the
header. 

[git-p4: depot-paths = "//depot/qt/main/": change = 118636]