p4i integration Integrate 117560 from 3 to main: Add GUIDs (Globally Unique IDentifiers) to all .pro files in Qt.
MSVC.NET uses GUIDs to do dependencies in a Solution file. If we don't have a GUID in the .pro file, we generate a new one every time. This would bring the Solution file out-of-sync, and mess up project dependencies.

When you create a new .pro file, please add a GUID. Thanks!

[git-p4: depot-paths = "//depot/qt/main/": change = 118719]