manual integration Integrate 118785 from 3 to main: Fix Intel C++ for Windows.
The problem is that existing Intel C++ configuration macros
were only triggered by Intel C++ for Linux. They were not
seen by Intel C++ for Windows. The reason is that Intel C++
for Windows disguises as Visual C++.

The solution I've chosen is to differentiate between Intel C++
for Windows and Intel C++ for Linux. I think these compilers
are different enough to warrant the duplication of macros.

[git-p4: depot-paths = "//depot/qt/main/": change = 118786]