p4i integration Integrate 120130 from 3 to main: add KDE patch task: 23151
Fixes flickering rubberbands in iconview and clears up in construction,
destruction of autoscroll timers. The problem was that we used the
contens position and called contains on a QRect constructed with
viewport dimensions. The contentspos the minute you scrolled would
always be bigger then the height of the viewport so it would never be
contained, the solution is ofcourse to use viewport position for this
test. Also respect the 50 margin, although I don't know if that is
really necessary.

[git-p4: depot-paths = "//depot/qt/main/": change = 120131]