p4i integration Integrate 120994 from 3 to main: backport change 109445 and change 106784 from main to 3.3... these changes were intended to be included after 3.2, but they were done before we decided to do 3.3:
change 109445:
make QMotif open its own display connect, and use this in QMotifWidget
and QMotifDialog when creating motif widgets.  This allows clipboard
operations to work between Qt and Motif widgets in the same
application.

Also, we remove the SubstructureRedirect mask from the standard QWidget
event mask... there is no need for it to be here, and it can actually
cause us problems.

change 106784:
Make click-to-focus work with Motif widgets that are QMotifWidget
children.

[git-p4: depot-paths = "//depot/qt/main/": change = 120995]