manually integrate change 121010:
handle select(2) errors more intelligently.  if select() fails because
it was interrupted by a signal, simply call select() again.  if
select() fails because of an invalid file descriptor, figure out which
socket notifier it belongs to and disable it.  this should prevent the
eventloop from getting into an unrecoverable state.

Fix verified on X11, but applies to QWS and Mac as well (needs to be
tested).

Reviewed by Jan Erik.

See task 34788.

[git-p4: depot-paths = "//depot/qt/main/": change = 121018]