p4i integration Integrate 121415 from 3 to main: Fix regression (it _must_ have been one) with keyboard selections in QListView (it works on Windows).
On X11 the ascii() value returned for Space and Ctrl-Space differs so
the old test didn't work and Ctrl-Space selections did not work. So
changed to use key() instead and moved it to the already present
Key_Space test. Also updated spacePressed and activate to be
emitted/called not only for Single selections but for Multi and
Extended as well.

autotest: 121414
review: mbm

[git-p4: depot-paths = "//depot/qt/main/": change = 121417]