p4i integration Integrate 122245 from 3 to main: Fix a bad behaviour introduced by change 95790 (which is a regression to Qt 3.1):
Enabling the read socket notifier in readBlock() and alike should
only be done if we are not in the sn_read() slot. This is because
before we emit the readyRead() signal in sn_read(), we disable the
read socket notifier.

This is done because a slot connected to the readyRead() signal
might cause event processing to happen and the sn_read() slot is
called again (if the socket notifier is enabled). This is a bad
thing, since we don't read any data from the socket to avoid
unwanted recursion in the user's code.

As a result if this happens, on Unix the socket notifier would fire
all the time, eating up CPU time. Even worse, on Windows the socket
notfier would only fire once and only again if more data arrives.
This might result in a dead lock situation, since data is not
reported when it is available.

So the socket notifier in the readBlock() and alike should only be
enabled if we are not in sn_read() (and if we are in sn_read(),
there is no need to enable it since after the readyRead() signal
was emitted we enable it anyway).

This was discovered by testing the QtSSLFilter class from Qt
Solutions where we show a modal dialog in a slot connected to the
readyRead() signal and where we were not able to receive the
welcome message until data arrived later on.

Reviewed by Andreas.

[git-p4: depot-paths = "//depot/qt/main/": change = 122246]