p4i integration Integrate 124030 from 3 to main: Fixed: QTable
Fix the problem described in Task 37169. Basically if someone created a
contenxt menu in a QTable, hit escape to cancel the menu, the mouse
press state would not be reset. Since we get the mouse button state in
QMouseEvent::state(), use that instead. I could not do this for
QTableHeader as for some wise reason use its buttonPressed value in
painting.

Approved by Trond and Marius B. M.

I guess I can throw a manual test into the mix.

[git-p4: depot-paths = "//depot/qt/main/": change = 124031]