Some small style changes.
First, make both the QuickDraw and CoreGraphics based styles inherit
from QCommonStyle instead of WindowsStyle (we already have enough
levels of indirection here).

Second, use GetThemeMetric as it seems like there is really no
replacement for it in HiThemes, we already share enums, why not a
function that returns an int. This makes checkboxes and radiobuttons
look okay.

[git-p4: depot-paths = "//depot/qt/main/": change = 124255]