p4i integration Integrate 124999 from 3 to main: 2 fixes for complex scripts:
qfontengine_x11: if a font has an invalid GPOS table
we ignore this fact and still mark it as supporting
open type if the GSUB table is fine. Fixes some issues
with some open type fonts.

qscriptengine_x11: Fix some of the issues with rendering
of khmer. At least the code doesn't crash anymore and
seems to be rendering the text partially correct.

Reviewed by Brad.

[git-p4: depot-paths = "//depot/qt/main/": change = 125000]