say hi to QLatin1String and QLatin1Char.
These classes can be used to wrap latin1 strings in the source
code, and provide optimised interaction with QString.

So 
if(str == QLatin1String("foo"))
can be a lot more efficient than
if(str == "foo")

The other advantage is that it will allow us to use latin1 strings
in Qt once we turn on QT_NO_CAST_FROM_ASCII and QT_NO_CAST_TO_ASCII
when compiling Qt itself.

[git-p4: depot-paths = "//depot/qt/main/": change = 126031]