p4i integration Integrate 128671 from 3 to main: Regression.
Workaround a problem on the Mac OS X where we could not quit anymore
when first loading assistant. We were able to do this before because of
flawed logic in qapplication_mac.cpp. Also this enabling and disabling
was superfluous because of the call qApp->processEvents(ExcludeUserInput);

Approved by Gunnar.

[git-p4: depot-paths = "//depot/qt/main/": change = 128672]