manually integrate # 128695
Xft2 support did in some cases skip certain (bitmap) fonts while building up
the database, so we thought the font wouldn't support scripts it actually did
support.

Added an FcCharSet containing the sample character for the script to the
requested pattern to get back a resaonable font (and not latin1 only).

Fixes tasks #38833 and 40893

Reviewed by Brad and Espen

[git-p4: depot-paths = "//depot/qt/main/": change = 128698]