manually integrate change 128715 from 3 to main:
Fix 2 window restore problem on Windows:

1. then the user minimizes the window using the titlebar icons, we
would get the correct event sequence: Hide, WindowStateChange,
ShowMinimized.  however, when restoring the window, we would not get a
Show event, only WindowStateChange and ShowNormal.

2. when restoring a minimized+maximized window, we would clear both the
minimized and maximized flags, while Windows only clears the minimized
flag.

the solution to both of these is to move the Show/Hide event code from
QETWidget::translateConfigEvent() into the SYSCOMMAND handler in
QtWndProc.  When minimizing, send the Hide event after setting the
WState_Minimized flag.  When restoring, we don't clear both the
minimized and maximized flags at the same time.  If the window is
minimized, we know we are restoring from the minimized state, so we
clear the flag and send the (previously missing) Show event; otherwise
we know that we are restoring from the maximized state.

See task 40828.

Done with Andy, reviewed by Marius S-O and Volker.

Manual test is in //depot/manualtests/main/qwidget/windowState/...

[git-p4: depot-paths = "//depot/qt/main/": change = 128716]