get rid of format references and use format groups instead that hide all the uglyness.
Give formats a pointer to their collection, so we can retrieve and set
a format group on them. Also added convenience method to get the table
and list format.

QtextFormat::toXxxFormat() now returns an invalid (and not an empty) format
when trying to convert something that is not of type Xxx.

This is just the first step, next will be to make QTextTable and
QTextList inherit from the groups. We'll then be able to get mostly
rid of the table and list manager.

[git-p4: depot-paths = "//depot/qt/main/": change = 130255]