Added Q_PRIVATE_SLOT macro.
The macro makes it possible to declare private slots in 
any slot section, with the symbol being declared in the private
class (the call will be to d->slot(...)). Rationale: we then do not need
to expose or export the symbol.

QLineEdit shows how to use it.

Important: if you use Q_PRIVATE_SLOT, you must include
the moc code at the end of the cpp file (like #include "moc_qlineedit.cpp").

[git-p4: depot-paths = "//depot/qt/main/": change = 130306]