It isn't pretty but I need to start somewhere, so I'm activating the new menubar code so that I can start the ports.
I have ported QWorkspace over to it (because it was the biggest abuser that I could see), it is mostly complete and seems to
mostly function. I added a couple gross functions for the controller widgets, this may need to be done differently, but this
will work for now. Additionally I took the opportunity to move QWorkspace over to the d/q pointer stuff since it would
need to be done at some point and this is just as good a time.

[git-p4: depot-paths = "//depot/qt/main/": change = 132201]