fixes task: 39834
adds check for d->startDragItem in contentsMouseMove so that for Single selection mode
when we start a click and move outside an selected icon and pass over it, we do not
start a new drag. (ie. only startDrag when startDragItem exists++)

dnd for Multi, Extended and NoSelection seems to work as before

[git-p4: depot-paths = "//depot/qt/main/": change = 133041]