change behavior on Multi selection mode for QIconView
now it only selects on mousepress, and deselects on mouserelease, this
means dnd becomes easier to do.  before you had to select and drag in
the same operation to be able to start a drag, if you first selected
then clicked and dragged you would deselect the clicked selected item
and no drag would start.

fixes task: 39863

[git-p4: depot-paths = "//depot/qt/main/": change = 133056]