first step towards introducing frames:
QTextFrame is a specialization of a QTextGroup. Opposed to
e.g. QTextList, the group is set on a char format, marking 
beginning and end of a frame.

Frames can be nested and give the document a way to express
hierachical structures.

Frames have to be set on the char and not on the blocks to
work consistently.

Lot's of work remains and all this is still not really working.

[git-p4: depot-paths = "//depot/qt/main/": change = 133814]